Fix DDD architecture issues in RuntimeErrorSage project

+++ src/RuntimeErrorSage.Core/Remediation/RemediationRegistry.cs
@@ -321,8 +321,8 @@ namespace RuntimeErrorSage.Application.Remediation
             try
             {
                 var strategies = _strategies.Values
-                    .Where(s => s.Priority >= minPriority)
-                    .OrderByDescending(s => s.Priority)
+                    .Where(s => (int)s.Priority >= minPriority)
+                    .OrderByDescending(s => (int)s.Priority)
                     .ToList();
 
                 _logger.LogDebug("Retrieved {Count} strategies with priority >= {MinPriority}", strategies.Count, minPriority);
